backend/svc-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
redis[hiredis]==5.2.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.20
python-dateutil==2.9.0
